api/requirements.txt
# Django and REST Framework
Django>=6.0.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Database
mysqlclient>=2.2.0

# Environment variables
python-decouple>=3.8

# Job Queue
celery>=5.4.0
redis>=5.0.0

# HTTP requests (for external API)
requests>=2.31.0

dj-database-url
whitenoise
gunicorn
psycopg2-binary
python-decouple

